vmm: support for driving the vmm from the outside
This commit does two things:
 - moves the vmm driving logic outside the vmm for more maneuverability,
 - completely decouples the vmm from the api_server, the vmm can now be
   controlled using whichever control method the external driver chooses.
