+++ api_server/src/request/machine_configuration.rs
@@ -6,10 +6,11 @@ use std::result;
 use futures::sync::oneshot;
 use hyper::{Method, Response, StatusCode};
 
+use super::VmmRequest;
 use http_service::json_response;
 use request::{GenerateHyperResponse, IntoParsedRequest, ParsedRequest};
 use vmm::vmm_config::machine_config::VmConfig;
-use vmm::{VmmAction, VmmRequest};
+use vmm::VmmAction;
 
 impl GenerateHyperResponse for VmConfig {
     fn generate_response(&self) -> Response {
