+++ api_server/src/request/net.rs
@@ -6,9 +6,10 @@ use std::result;
 use futures::sync::oneshot;
 use hyper::Method;
 
+use super::VmmRequest;
 use request::{IntoParsedRequest, ParsedRequest};
 use vmm::vmm_config::net::{NetworkInterfaceConfig, NetworkInterfaceUpdateConfig};
-use vmm::{VmmAction, VmmRequest};
+use vmm::VmmAction;
 
 impl IntoParsedRequest for NetworkInterfaceConfig {
     fn into_parsed_request(
