+++ api_server/src/request/vsock.rs
@@ -6,9 +6,10 @@ use std::result;
 use futures::sync::oneshot;
 use hyper::Method;
 
+use super::VmmRequest;
 use request::{IntoParsedRequest, ParsedRequest};
 use vmm::vmm_config::vsock::VsockDeviceConfig;
-use vmm::{VmmAction, VmmRequest};
+use vmm::VmmAction;
 
 impl IntoParsedRequest for VsockDeviceConfig {
     fn into_parsed_request(
